requirements.txt
# Backend B2B Coworking API
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
psycopg2-binary>=2.9.0
pydantic-settings>=2.0.0
email-validator>=2.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
pandas>=2.0.0
openpyxl>=3.1.0
python-multipart>=0.0.9
